env variable for API url. Edit bio works with API
