--- src/components/RecentPosts/_RecentPosts.scss
@@ -1,14 +1,0 @@
-.RecentPosts-container {
-  display: flex;
-  flex-direction: column;
-  @extend %ui-content-container;
-  .RecentPosts-actions {
-    margin-left: 0.5rem;
-    button {
-      margin: 0;
-    }
-    button:nth-of-type(1) {
-      margin-right: 0.5rem;
-    }
-  }
-}
