+++ src/components/_components_index.scss
@@ -24,8 +24,6 @@
 @import "./Messages/forms/SendMessageForm";
 @import "./Messages/forms/ReplyToMessageForm";
 
-@import "./RecentPosts/RecentPosts";
-
 @import "./Group/GroupInfo/GroupInfo";
 @import "./Group//GroupMembers/GroupMembers";
 @import "./Group/GroupNewPosts/GroupNewPosts";
