+++ src/containers/GroupsListPage/GroupsListPage.js
@@ -10,7 +10,7 @@ export const GroupsListPage = (props) => {
   const [groupsForPage, setGroupsForPage] = useState(null);
   useEffect(() => {
     axios
-      .get("http://localhost:5000/api/v1/groups")
+      .get(`${process.env.REACT_APP_API_URL}/groups`)
       .then((res) => {
         setGroupsForPage(res.data);
         setPageStatus({
