+++ src/containers/RecentPostsPage/_RecentPostsPage.scss
@@ -1,3 +1,17 @@
 .RecentPostsPage-container {
   @extend %page-center-content;
+  .RecentPosts-container {
+    display: flex;
+    flex-direction: column;
+    @extend %ui-content-container;
+    .RecentPosts-actions {
+      margin-left: 0.5rem;
+      button {
+        margin: 0;
+      }
+      button:nth-of-type(1) {
+        margin-right: 0.5rem;
+      }
+    }
+  }
 }
